Refactor TradeDashboardController to cast realized PNL and entry price to float. Update trade_dashboard view to improve layout and add invested amount and close reason with emojis for better clarity. Adjust grid columns from 5 to 6 and enhance text styling for better readability

+++ app/Http/Controllers/TradeDashboardController.php
@@ -181,8 +181,8 @@ class TradeDashboardController extends Controller
                 ->limit(10)
                 ->get()
                 ->map(function ($pos) {
-                    $pnl = $pos->realized_pnl ?? 0;
-                    $entryPrice = $pos->entry_price;
+                    $pnl = (float) ($pos->realized_pnl ?? 0);
+                    $entryPrice = (float) $pos->entry_price;
 
                     // Calculate PNL percentage with leverage
                     $pnlPercent = 0;
